requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-reversion==3.0.4
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
Markdown==3.1.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.10.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
